Add /new endpoint and hashing
Adds a /new endpoint to allow for creating short URLs. Adds hashing to
ensures that only one short URL is created for each long URL. This is
implemented using the DJB2 hashing algorithm. The urls table is indexed
on hash to make lookup on potential insertion faster. Hash and url are
searched for before insert to check whether or not a new entry needs to
be created.